feature/num_study_phenotype_aqls Implement an integration test to persist/query COMPOSITION based on openEHR confirmed COVID-19 infection report.v0 template
A flat-DTO/entity was generated using the generator tool of the SDK and a COMPOSITION instance based on this DTO is pushed to EhrBase.
This COMPOSITION is then queried via AQL.

DefaultRestClient.java is modified to add the capability to use proxy servers, which is helpful during development, when using tools such as Fiddler, Postman or Charles.